uv-metadata/pyproject.toml
[project]
name = "sb-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=11.2.1",
]
